requirements.txt
fastapi
uvicorn
pandas
supabase
google-generativeai
python-dotenv